app/personal/serializers.py
# personal/serializers.py
from rest_framework import serializers
from .models import Profesor, Alumno

class ProfesorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Profesor
        fields = '__all__'

class AlumnoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Alumno
        fields = '__all__'