requirements.txt
Django>=4.2
psycopg2-binary>=2.9
google-generativeai
python-dotenv
djangorestframework
psycopg2-binary
numpy
gunicorn
scikit-learn
